### 2023-01-11 d9f0e08
Improve comparison between linear subspaces
