### 2023-01-11 4a2e8c1
More robust independent subspace comparison
