### 2023-01-11 c997672
Fix orthonormalization
