### 2023-01-12 #5064 @lan496
@danielskatz Thank you for starting the pre-review process.
This package is a bit interdisciplinary: the implementation is related to (elemental) computational group theory or representation theory (Track 7 CSISM), but the main application is for condensed matter physics and materials science (Track 2 BCM). I am not sure which tracks are appropriate.

Here are potential reviewers from the list (specialized in computational physics, materials science, or crystallography):
bjmorgan
jarvist
mkhorton
mdavezac
ml-evs
hungpham2017
