+++ develop/_sources/examples/lattice_vibration.ipynb.txt
@@ -1,0 +1,412 @@
+{
+ "cells": [
+  {
+   "cell_type": "markdown",
+   "metadata": {},
+   "source": [
+    "# Application of space-group irreps to lattice vibration\n",
+    "\n",
+    "This example shows how to prepare representation matrix of atomic displacements by space group, and block-diagonalize dynamical matrix for harmonic lattice vibration only from symmetry."
+   ]
+  },
+  {
+   "cell_type": "markdown",
+   "metadata": {},
+   "source": [
+    "## Import modules"
+   ]
+  },
+  {
+   "cell_type": "markdown",
+   "metadata": {},
+   "source": []
+  },
+  {
+   "cell_type": "code",
+   "execution_count": 1,
+   "metadata": {},
+   "outputs": [
+    {
+     "name": "stdout",
+     "output_type": "stream",
+     "text": [
+      "spgrep==0.3.2.dev5+ga44d231\n"
+     ]
+    }
+   ],
+   "source": [
+    "from pathlib import Path\n",
+    "\n",
+    "import phonopy\n",
+    "import numpy as np\n",
+    "import matplotlib.pyplot as plt\n",
+    "import seaborn as sns\n",
+    "\n",
+    "import spgrep\n",
+    "from spgrep import get_spacegroup_irreps\n",
+    "from spgrep.representation import check_spacegroup_representation, project_to_irrep\n",
+    "\n",
+    "print(f\"spgrep=={spgrep.__version__}\")"
+   ]
+  },
+  {
+   "cell_type": "code",
+   "execution_count": 2,
+   "metadata": {},
+   "outputs": [],
+   "source": [
+    "sns.set_context(\"poster\")"
+   ]
+  },
+  {
+   "cell_type": "markdown",
+   "metadata": {},
+   "source": [
+    "## Prepare space-group operations and representation\n",
+    "\n",
+    "We consider lattice vibration of Perovskite structure ($Pm\\overline{3}m$) at X point."
+   ]
+  },
+  {
+   "cell_type": "code",
+   "execution_count": 3,
+   "metadata": {},
+   "outputs": [],
+   "source": [
+    "# Perovskite structure: Pm-3m (No. 221)\n",
+    "a = 3.986\n",
+    "lattice = np.array(\n",
+    "    [\n",
+    "        [a, 0, 0],\n",
+    "        [0, a, 0],\n",
+    "        [0, 0, a],\n",
+    "    ]\n",
+    ")\n",
+    "positions = np.array(\n",
+    "    [\n",
+    "        [0, 0.5, 0.5],  # O(3c)\n",
+    "        [0.5, 0, 0.5],  # O(3c)\n",
+    "        [0.5, 0.5, 0],  # O(3c)\n",
[…]
