+++ develop/_sources/examples/symmetry_adapted_tensor.ipynb.txt
@@ -1,0 +1,402 @@
+{
+ "cells": [
+  {
+   "cell_type": "markdown",
+   "metadata": {},
+   "source": [
+    "# Symmetry-adapted tensor\n",
+    "\n",
+    "This example shows how to calculate irreps of crystallographic point groups and construct projection operator on identity representation.\n",
+    "We apply the projection operator to obtaining symmetry-adapted tensor (e.g. elastic constant)."
+   ]
+  },
+  {
+   "cell_type": "markdown",
+   "metadata": {},
+   "source": [
+    "## Import modules"
+   ]
+  },
+  {
+   "cell_type": "code",
+   "execution_count": 1,
+   "metadata": {},
+   "outputs": [
+    {
+     "name": "stdout",
+     "output_type": "stream",
+     "text": [
+      "spgrep==0.3.2.dev5+ga44d231\n"
+     ]
+    }
+   ],
+   "source": [
+    "from __future__ import annotations\n",
+    "\n",
+    "import matplotlib.pyplot as plt\n",
+    "import numpy as np\n",
+    "import seaborn as sns\n",
+    "from spglib import get_symmetry_from_database\n",
+    "\n",
+    "import spgrep\n",
+    "from spgrep import get_crystallographic_pointgroup_irreps_from_symmetry\n",
+    "from spgrep.group import get_cayley_table\n",
+    "from spgrep.representation import is_representation, project_to_irrep\n",
+    "from spgrep.tensor import get_symmetry_adapted_tensors, apply_intrinsic_symmetry\n",
+    "\n",
+    "print(f\"spgrep=={spgrep.__version__}\")"
+   ]
+  },
+  {
+   "cell_type": "code",
+   "execution_count": 2,
+   "metadata": {},
+   "outputs": [],
+   "source": [
+    "sns.set_context(\"poster\")"
+   ]
+  },
+  {
+   "cell_type": "markdown",
+   "metadata": {},
+   "source": [
+    "## Prepare point-group operations and representation\n",
+    "\n",
+    "We consider crystallographic point group $m\\overline{3}m$."
+   ]
+  },
+  {
+   "cell_type": "code",
+   "execution_count": 3,
+   "metadata": {},
+   "outputs": [],
+   "source": [
+    "# Pm-3m (No. 221)\n",
+    "symmetry = get_symmetry_from_database(hall_number=517)\n",
[…]
