+++ develop/_sources/formulation/algorithm/irreps_from_chain.md.txt
@@ -1,0 +1,106 @@
+# On-the-fly irreps generation from solvable group chain
+
+This page describes an algorithm to generate irreps for solvable groups {cite}`Neto:a09740`.
+
+## Subduced and induced representations for solvable group
+
+Let $H$ be invariant subgroup of $G$ such that $G/H = \mathbb{Z}_{p}$ ($p$ is prime and $r^{p}=e$),
+
+$$
+  G = \coprod_{m=0}^{p-1} r^{m} H.
+$$ (decomp)
+
+Consider projective irrep $\Delta$ with factor system $\mu$ for $H$,
+$$
+  S \psi_{j} = \sum_{i=1}^{d} \psi_{i} \Delta(S)_{ij} \quad (R \in H, i = 1, \dots, d).
+$$
+The following representations are also irrep over $H$:
+$$
+  \mathbf{\Delta}^{(m)}(S)
+    &:= \frac{ \mu(S, r^{m}) }{ \mu(r^{m}, S_{m}) } \mathbf{\Delta}(S_{m})
+    \quad (S \in H, m = 0, \dots, p-1) \\
+  S_{m}
+    &:= r^{-m} S r^{m} (\in H)
+$$
+There are two cases for {math}`\{ \Delta^{(m)} \}_{m=0, \dots, p-1}`:
+1. They are mutually inequivalent: {math}`\Delta \uparrow G \downarrow H \cong \sum_{m=0}^{p-1} \Delta^{(m)}`
+2. They are all equivalent: {math}`\Delta \uparrow G \downarrow H \cong p \Delta`
+
+### Case-1: conjugated irreps are mutually inequivalent
+
+For case-1, the induced representation $\Delta \uparrow G$ is irrep [^footnote1] [^footnote2]:
+$$
+  \Delta \uparrow G (r^{m} S)_{i:, j:}
+    &= \mathbb{I}[i \equiv j + m] \frac{ \mu(r^{m}S, r^{j}) }{ \mu(r^{i}, S_{j} ) } \mathbf{\Delta}( S_{j} )
+    \quad (S \in H).\\
+  \Delta \uparrow G (S)
+    &= \begin{pmatrix}
+      \mathbf{\Delta}^{(0)}(S) & & & \\
+      & \mathbf{\Delta}^{(1)}(S) & & \\
+      & & \ddots & \\
+      & & & \mathbf{\Delta}^{(p-1)}(S)
+    \end{pmatrix}
+    \quad (S \in H) \\
+  \Delta \uparrow G (r)
+    &= \begin{pmatrix}
+      \mathbf{0} & \mathbf{0} & \mathbf{0} & \cdots & \mathbf{0} & \frac{\mu(r, r^{p-1})}{\mu(E,E)} \mathbf{1} \\
+      \frac{\mu(r, E)}{\mu(r,E)} \mathbf{1} & \mathbf{0} & \mathbf{0} & \cdots & \mathbf{0} & \mathbf{0} \\
+      \mathbf{0} & \frac{\mu(r, r)}{\mu(r^{2},E)} \mathbf{1} & \mathbf{0} & \cdots & \mathbf{0} & \mathbf{0} \\
+      \vdots     & \vdots     & \vdots     &        & \vdots     & \vdots \\
+      \mathbf{0} & \mathbf{0} & \mathbf{0} & \cdots & \frac{\mu(r, r^{p-2})}{\mu(r^{p-1},E)} \mathbf{1} & \mathbf{0}
+    \end{pmatrix}.
+$$
+
+[^footnote1]: This follows from Frobenius reciprocity theorem. In this case, $\Delta$ appears only once in $\Gamma \downarrow H$.
[…]
