### 2023-01-23 507bc75
[pre-commit.ci] pre-commit autoupdate
updates:
- [github.com/pycqa/pydocstyle: 6.2.3 → 6.3.0](https://github.com/pycqa/pydocstyle/compare/6.2.3...6.3.0)
