+++ .pre-commit-config.yaml
@@ -47,7 +47,7 @@ repos:
       - id: pyupgrade
   # Docstring
   - repo: https://github.com/pycqa/pydocstyle
-    rev: 6.2.3
+    rev: 6.3.0
     hooks:
     - id: pydocstyle
       additional_dependencies: ["toml"]
