### 2023-01-25 077dc01
Add comments of Seitz symbols for point groups
