+++ src/spgrep/pointgroup.py
@@ -9,12 +9,13 @@ from spgrep.utils import NDArrayInt, ndarray2d_to_integer_tuple
 # List of point groups after applying transformation matrices given by `spglib.get_pointgroup`
 # See https://github.com/spglib/spglib/issues/164
 # Operations are ordered as same as Table 3.2.3.2 of ITA (2016).
+# fmt: off
 pg_dataset = {
-    "1": [[((1, 0, 0), (0, 1, 0), (0, 0, 1))]],
+    "1": [[((1, 0, 0), (0, 1, 0), (0, 0, 1))]],  # 0: 1
     "-1": [
         [
-            ((1, 0, 0), (0, 1, 0), (0, 0, 1)),
-            ((-1, 0, 0), (0, -1, 0), (0, 0, -1)),
+            ((1, 0, 0), (0, 1, 0), (0, 0, 1)),  # 0: 1
+            ((-1, 0, 0), (0, -1, 0), (0, 0, -1)),  # 1: -1
         ]
     ],
     "2": [
@@ -25,8 +26,8 @@ pg_dataset = {
         ],
         # unique axis-b
         [
-            ((1, 0, 0), (0, 1, 0), (0, 0, 1)),
-            ((-1, 0, 0), (0, 1, 0), (0, 0, -1)),
+            ((1, 0, 0), (0, 1, 0), (0, 0, 1)),  # 0: 1
+            ((-1, 0, 0), (0, 1, 0), (0, 0, -1)),  # 1: 2
         ],
         # unique axis-c
         [
@@ -42,8 +43,8 @@ pg_dataset = {
         ],
         # unique axis-b
         [
-            ((1, 0, 0), (0, 1, 0), (0, 0, 1)),
-            ((1, 0, 0), (0, -1, 0), (0, 0, 1)),
+            ((1, 0, 0), (0, 1, 0), (0, 0, 1)),  # 0: 1
+            ((1, 0, 0), (0, -1, 0), (0, 0, 1)),  # 1: m
         ],
         # unique axis-c
         [
@@ -54,18 +55,18 @@ pg_dataset = {
     "2/m": [
         # unique axis-b
         [
-            ((1, 0, 0), (0, 1, 0), (0, 0, 1)),
-            ((-1, 0, 0), (0, 1, 0), (0, 0, -1)),
-            ((-1, 0, 0), (0, -1, 0), (0, 0, -1)),
-            ((1, 0, 0), (0, -1, 0), (0, 0, 1)),
+            ((1, 0, 0), (0, 1, 0), (0, 0, 1)),  # 0: 1
+            ((-1, 0, 0), (0, 1, 0), (0, 0, -1)),  # 1: 2
+            ((-1, 0, 0), (0, -1, 0), (0, 0, -1)),  # 2: -1
+            ((1, 0, 0), (0, -1, 0), (0, 0, 1)),  # 3: m
         ]
     ],
     "222": [
         [
-            ((1, 0, 0), (0, 1, 0), (0, 0, 1)),
-            ((-1, 0, 0), (0, -1, 0), (0, 0, 1)),
-            ((-1, 0, 0), (0, 1, 0), (0, 0, -1)),
-            ((1, 0, 0), (0, -1, 0), (0, 0, -1)),
+            ((1, 0, 0), (0, 1, 0), (0, 0, 1)),  # 0: 1
+            ((-1, 0, 0), (0, -1, 0), (0, 0, 1)),  # 1: 2_001
[…]
