+++ setup.cfg
@@ -1,7 +1,7 @@
 [flake8]
 max-line-length = 99
 select = C,E,F,W,B,B950
-ignore = E203, E231, E501, W503, W605
+ignore = E203, E231, E262, E501, W503, W605
 
 [coverage:run]
 omit = tests/
