### 2023-01-30 a852787
[pre-commit.ci] pre-commit autoupdate
updates:
- [github.com/pycqa/isort: 5.11.4 → 5.12.0](https://github.com/pycqa/isort/compare/5.11.4...5.12.0)
