+++ .pre-commit-config.yaml
@@ -29,7 +29,7 @@ repos:
       exclude: ^docs/
   # isort
   - repo: https://github.com/pycqa/isort
-    rev: 5.11.4
+    rev: 5.12.0
     hooks:
       - id: isort
         name: isort (python)
