+++ src/spgrep/pointgroup.py
@@ -355,8 +355,8 @@ pg_dataset = {
             ((-1, 1, 0), (0, 1, 0), (0, 0, 1)),   #  7: m_100
             ((1, 0, 0), (1, -1, 0), (0, 0, 1)),   #  8: m_010
             ((0, 1, 0), (1, 0, 0), (0, 0, 1)),    #  9: m_1-10
-            ((1, -1, 0), (0, -1, 0), (0, 0, 1)),  # 10: m_110
-            ((-1, 0, 0), (-1, 1, 0), (0, 0, 1)),  # 11: m_100
+            ((1, -1, 0), (0, -1, 0), (0, 0, 1)),  # 10: m_120
+            ((-1, 0, 0), (-1, 1, 0), (0, 0, 1)),  # 11: m_210
         ]
     ],
     "-6m2": [
