+++ src/spgrep/pointgroup.py
@@ -272,8 +272,8 @@ pg_dataset = {
             ((0, 1, 0), (-1, 1, 0), (0, 0, -1)),   #  7: -3^+
             ((1, -1, 0), (1, 0, 0), (0, 0, -1)),   #  8: -3^-
             ((0, 1, 0), (1, 0, 0), (0, 0, 1)),     #  9: m_1-10
-            ((1, -1, 0), (0, -1, 0), (0, 0, 1)),   # 10: m_010
-            ((-1, 0, 0), (-1, 1, 0), (0, 0, 1)),   # 11: m_100
+            ((1, -1, 0), (0, -1, 0), (0, 0, 1)),   # 10: m_120
+            ((-1, 0, 0), (-1, 1, 0), (0, 0, 1)),   # 11: m_210
         ],
         # -3m1
         [
