### 2023-02-20 ffebb35
[pre-commit.ci] pre-commit autoupdate
updates:
- [github.com/pre-commit/mirrors-mypy: v1.0.0 → v1.0.1](https://github.com/pre-commit/mirrors-mypy/compare/v1.0.0...v1.0.1)
- [github.com/nbQA-dev/nbQA: 1.6.1 → 1.6.3](https://github.com/nbQA-dev/nbQA/compare/1.6.1...1.6.3)
