+++ .pre-commit-config.yaml
@@ -23,7 +23,7 @@ repos:
       exclude: ^docs/
   # type annotation
   - repo: https://github.com/pre-commit/mirrors-mypy
-    rev: v1.0.0
+    rev: v1.0.1
     hooks:
     - id: mypy
       exclude: ^docs/
@@ -54,6 +54,6 @@ repos:
       files: ^src/spgrep/
   # Notebook
   - repo: https://github.com/nbQA-dev/nbQA
-    rev: 1.6.1
+    rev: 1.6.3
     hooks:
       - id: nbqa-black
