### 2023-03-06 #5064 @drvinceknight
@bjmorgan @jarvist  @mkhortonI am pinging you to ask if you would be able to assist with reviewing this submissions to JOSS (The Journal of Open Source Software).

JOSS publishes articles about open source research software. The submission I'd like you to review is titled: “spgrep: On-the-fly generator of space-group irreducible representations”. You can find more information at the top of this Github issue.
