### 2023-03-20 #5269 @editorialbot
[REVIEW]: spgrep: On-the-fly generator of space-group irreducible representations
**Submitting author:** @lan496 (<a href="http://orcid.org/0000-0002-5907-2549">Kohei Shinohara</a>)
**Repository:** https://github.com/spglib/spgrep
**Branch with paper.md** (empty if default branch):
**Version:** v0.3.4
