### 2023-04-03 ab24e63
Update sphinx requirement from <4.3 to <6.2
Updates the requirements on [sphinx](https://github.com/sphinx-doc/sphinx) to permit the latest version.
- [Release notes](https://github.com/sphinx-doc/sphinx/releases)
- [Changelog](https://github.com/sphinx-doc/sphinx/blob/master/CHANGES)
- [Commits](https://github.com/sphinx-doc/sphinx/compare/0.1.61611...v6.1.3)

---
updated-dependencies:
- dependency-name: sphinx
  dependency-type: direct:development
...

Signed-off-by: dependabot[bot] <support@github.com>
