+++ setup.py
@@ -43,7 +43,7 @@ EXTRAS = {
     "docs": [
         # We cannot update sphinx to >6 until docutils==0.20 is released: https://github.com/mcmtroffaes/sphinxcontrib-bibtex/issues/322
         # Another issue requires sphinx<4.3 (fixed in pydata-sphinx-theme>=0.8.0): https://github.com/pydata/pydata-sphinx-theme/pull/509
-        "docutils<=0.17",
+        "docutils<=0.19",
         "sphinx<4.3",
         "sphinx-autobuild",
         "nbsphinx",
