### 2023-04-03 38257b2
[pre-commit.ci] pre-commit autoupdate
updates:
- [github.com/psf/black: 23.1.0 → 23.3.0](https://github.com/psf/black/compare/23.1.0...23.3.0)
- [github.com/nbQA-dev/nbQA: 1.6.4 → 1.7.0](https://github.com/nbQA-dev/nbQA/compare/1.6.4...1.7.0)
