### 2023-04-20 #5269 @mdavezac
spgrep gives users the ability to compute irreps from a minimal input set consistent with the inputs and outputs common in atomistic simulations. As such, it makes it easier to created simulation that mesh atomistic codes with group-theoretic methods. The paper is well written, the code is well tested and documented. I am happy to recommend it for publication in JOSS.
