### 2023-05-01 bb02e71
Update sphinx requirement from <4.3 to <7.1
Updates the requirements on [sphinx](https://github.com/sphinx-doc/sphinx) to permit the latest version.
- [Release notes](https://github.com/sphinx-doc/sphinx/releases)
- [Changelog](https://github.com/sphinx-doc/sphinx/blob/master/CHANGES)
- [Commits](https://github.com/sphinx-doc/sphinx/compare/v0.1.61611...v7.0.0)

---
updated-dependencies:
- dependency-name: sphinx
  dependency-type: direct:development
...

Signed-off-by: dependabot[bot] <support@github.com>
