### 2023-05-01 spglib/spgrep#77 @ml-evs
Missing data file in lattice vibration example
Hi there, I'm reviewing this repository for JOSS. Everything else looks great to me, but I am unable to run `examples/lattice_vibration.py` as the file `phonopy_mp-2998.yaml.xz` is missing, and I couldn't find it in the raw data from http://phonondb.mtl.kyoto-u.ac.jp/ph20180417/d002/mp-2998.html. Is this an oversight, or am I missing something?
