### 2023-05-01 #5269 @ml-evs
@drvinceknight I have completed my review and I am happy to recommend spgrep for publication in JOSS, however I just have one minor outstanding issue with running one of the examples: https://github.com/spglib/spgrep/issues/77 I don't see this as blocking and it can be addressed in the final archival of the repository ready for publication.

The package is well-motivated, well-developed and well-documented, great job!
