### 2023-05-02 spglib/spgrep#77 @lan496
@ml-evs Thank you for examining the example script. I have fixed the location for phonopy data `phonopy_mp-2998.yaml.xz`. Can you check if the issue is solved at the HEAD of develop branch (https://github.com/spglib/spgrep/commit/161fd31ea469702083934cecbcf50fb662d934ab)?
Actually, xz-format phonopy datum are stored in <http://phonondb.mtl.kyoto-u.ac.jp/xz-files/> in phonondb.
