### 2023-05-02 spglib/spgrep#77 @ml-evs
Thanks, that works for me now! Indeed, I was running the script rather than the Jupyter notebook, which probably always worked.
