+++ .pre-commit-config.yaml
@@ -42,7 +42,7 @@ repos:
         types: [pyi]
   # Upgrade syntax
   - repo: https://github.com/asottile/pyupgrade
-    rev: v3.3.2
+    rev: v3.4.0
     hooks:
       - id: pyupgrade
   # Docstring
