### 2023-05-11 4cd5042
Fix bibtex entry
