+++ docs/paper/paper.bib
@@ -250,7 +250,7 @@
   volume = {221},
   number = {1},
   journal = {Zeitschrift f\"{u}r Kristallographie - Crystalline Materials},
-  doi = {doi:10.1524/zkri.2006.221.1.15},
+  doi = {10.1524/zkri.2006.221.1.15},
   year = {2006},
   lastchecked = {2022-12-19}
 }
