### 2023-05-15 9ebc095
Bump phonopy from 2.18.0 to 2.19.0
Bumps [phonopy](https://phonopy.github.io/phonopy/) from 2.18.0 to 2.19.0.

---
updated-dependencies:
- dependency-name: phonopy
  dependency-type: direct:development
  update-type: version-update:semver-minor
...

Signed-off-by: dependabot[bot] <support@github.com>
