+++ develop/_sources/api/api.md
@@ -1,14 +1,0 @@
-# API Reference
-
-```{toctree}
-    Core functions <api_core>
-    Group <api_group>
-    Point group <api_pointgroup>
-    Transformation <api_transformation>
-    Representation <api_representation>
-    Irreps <api_irreps>
-    Spinor <api_spinor>
-    Co-representation <api_corep>
-    Tensor <api_tensor>
-    Utility functions <api_utils>
-```
