+++ develop/_sources/formulation/algorithm/irreps_from_regular.md
@@ -1,170 +1,0 @@
-# On-the-fly irreps generation from regular representation
-
-This page describes an algorithm for generating all irreps by random matrix.
-
-## Regular representation
-
-Given a finite group $G = \{ R_{k} \}_{k=1}^{|G|}$, the regular representation of $G$ is defined as
-$$
-  \Gamma^{(\mathrm{reg})}(G_{k})_{ij}
-    = \delta( G_{i}, G_{k} G_{j} ),
-$$
-where $\delta(\cdot, \cdot)$ takes one iff two elements are equal.
-
-For any factor system $\mu$, the following representation holds the orthogonality theorem,
-$$
-  \Delta^{(\mathrm{reg})}(G_{k})_{ij}
-    = \delta(G_{i}, G_{k}G_{j}) \mu(G_{k}, G_{j}).
-$$
-We refer this representation to projective regular representation.
-
-The projective regular representation is a unitary representation when absolute of its factor system is one:
-$$
-  \sum_{l} \Delta^{(\mathrm{reg})}(G_{k})_{il} \Delta^{(\mathrm{reg})}(G_{k})_{jl}^{\ast}
-  = \delta_{ij} \left| \mu(G_{k}, G_{k}^{-1}G_{i}) \right|^{2}.
-$$
-The factor system defined in {ref}`space_group_irreps` satisfies this condition.
-
-## Obtaining all irreps from (projective) regular representation
-
-Refs. {cite}`THOMAS201776,doi:10.1137/090779966,Po2017`
-
-For a finite group $G$ and its (projective) regular representation $\Gamma^{(\mathrm{reg})}$,
-$$
-  \tilde{\mathbf{H}}
-    = \sum_{R \in G} \mathbf{\Gamma}^{(\mathrm{reg})}(R) \mathbf{H} \mathbf{\Gamma}^{(\mathrm{reg})}(R^{-1})
-$$
-is also Hermite and commute with $\Gamma^{(\mathrm{reg})}$ [^footnote0], where $\mathbf{H}$ is any Hermite matrix.
-Consider the spectral decomposition of $\tilde{\mathbf{H}}$,
-$$
-  \tilde{\mathbf{H}} = \sum_{\lambda} \lambda \sum_{n=1}^{ d_{\lambda} } \mathbf{v}_{\lambda n} \mathbf{v}_{\lambda n}^{\dagger} \\
-  \lambda \in \mathbb{R}, \mathbf{v}_{\lambda n}^{\dagger} \mathbf{v}_{\lambda' n'} = \delta_{\lambda \lambda'} \delta_{n n'}.
-$$
-Then, the regular representation is block-diagonalized with $\mathbf{V} = ( \mathbf{v}_{\lambda_{1} 1} \dots \mathbf{v}_{\lambda_{1} d_{\lambda_{1}}} \dots )$,
-$$
-  \left[ \mathbf{V}^{\dagger} \mathbf{\Gamma}^{(\mathrm{reg})}(R) \mathbf{V} \right]_{\lambda n, \lambda' n'} (\lambda - \lambda') = 0
-$$
-
-[^footnote0]: The derivation is as follows:
-    ```{math}
-      \mathbf{\Delta}^{(\mathrm{reg})}(S) \mathbf{U} \mathbf{\Delta}^{(\mathrm{reg})}(S)^{-1}
[…]
