+++ develop/_sources/api/api_utils.md
@@ -1,5 +1,0 @@
-# Utility functions
-
-```{eval-rst}
-    .. autofunction:: spgrep.utils.grassmann_distance
-```
