+++ develop/_sources/api/api_corep.md
@@ -1,9 +1,0 @@
-# Co-representation for spinor
-
-```{eval-rst}
-    .. autofunction:: spgrep.corep.enumerate_spinor_small_corepresentations
-```
-
-```{eval-rst}
-    .. autofunction:: spgrep.corep.get_corep_spinor_factor_system
-```

[…]
