### 2023-05-17 #5269 @lan496
Thank you for your reply!  I have finished:
- tagged release: v0.3.4 https://github.com/spglib/spgrep/releases/tag/v0.3.4
- archive (zenodo): https://zenodo.org/record/7946899
- doi: 10.5281/zenodo.7946899
