### 2023-05-22 92545a9
Bump phonopy from 2.19.0 to 2.19.1
Bumps [phonopy](https://phonopy.github.io/phonopy/) from 2.19.0 to 2.19.1.

---
updated-dependencies:
- dependency-name: phonopy
  dependency-type: direct:development
  update-type: version-update:semver-patch
...

Signed-off-by: dependabot[bot] <support@github.com>
