+++ develop/_sources/formulation/irreps/spacegroup_irreps.md
@@ -1,0 +1,157 @@
+(space_group_irreps)=
+# Irreps of space group
+
+This page formulates little group, small representation, and irreducible representations (irreps) of space groups.
+
+## Reciprocal lattice and transformation
+
+Let $\mathbf{A} = (\mathbf{a}_{1}, \mathbf{a}_{2}, \mathbf{a}_{3})$ be basis of $L_{\mathcal{T}}$, which is lattice of translational group ${\mathcal{T}}$.
+We choose basis of the (crystallographic) reciprocal vectors as
+$$
+  \mathbf{B} &= (\mathbf{b}_{1}, \mathbf{b}_{2}, \mathbf{b}_{3}) = A^{-\top} \\
+  \mathbf{b}_{i} \cdot \mathbf{a}_{j} &= \delta_{ij}.
+$$
+
+When basis vectors $\mathbf{A}$ is transformed to $\mathbf{A}' := \mathbf{AP}$,
+- its dual basis vectors in reciprocal space is transformed from $\mathbf{B}=\mathbf{A}^{-\top}$ to $\mathbf{B}' := \mathbf{B}\mathbf{P}^{-\top}$.
+- symmetry operation $(\mathbf{R}, \mathbf{v})$ is transformed to
+    ```{math}
+        (\mathbf{P}, \mathbf{0})^{-1} (\mathbf{R}, \mathbf{v}) (\mathbf{P}, \mathbf{0})
+            = (\mathbf{P}^{-1}\mathbf{RP}, \mathbf{P}^{-1}\mathbf{v}).
+    ```
+- Coefficients of reciprocal vector $\mathbf{k} = 2 \pi \mathbf{B} \mathbf{k}_{f}$ is transformed to $\mathbf{k}_{f}' := \mathbf{P}^{\top} \mathbf{k}_{f}$.
+
+## Irreps of translation subgroup
+
+Consider the following irreps of translation subgroup $\mathcal{T}$ of space group $\mathcal{G}$,
+$$
+  \Gamma^{(\mathbf{k})} ((\mathbf{E}, \mathbf{t}))
+  = \exp \left( -i \mathbf{k} \cdot \mathbf{t} \right)
+  \quad ( \mathbf{t} \in L_{\mathcal{T}} ),
+$$
+where $L_{\mathcal{T}}$ is a lattice formed by $\mathcal{T}$.
+Let $L_{\mathcal{T}}^{\ast}$ be a reciprocal lattice of $L_{\mathcal{T}}$.
+We can confine $\mathbf{k}$ within or the surface of a primitive cell of the reciprocal space because, for $\mathbf{g} \in L_{\mathcal{T}}^{\ast}$, $\Gamma^{(\mathbf{k})} = \Gamma^{(\mathbf{k} + \mathbf{g})}$.
+Any Bloch function with $\mathbf{k}$
+$$
+  \Psi_{\mathbf{k}}(\mathbf{r}) = e^{ i \mathbf{k} \cdot \mathbf{r} } u_{\mathbf{k}}(\mathbf{r}),
+$$
+where $u_{\mathbf{k}}(\mathbf{r})$ is periodic with $L_{\mathcal{T}}$, can be taken as (one-dimensional) basis of $\Gamma^{\mathbf{k}}$:
+$$
+  (\mathbf{E}, \mathbf{t}) \Psi_{\mathbf{k}}(\mathbf{r})
[…]
