+++ develop/_sources/api/api_representation.md
@@ -1,0 +1,41 @@
+# Representation
+
+```{eval-rst}
+    .. autofunction:: spgrep.representation.get_regular_representation
+```
+
+```{eval-rst}
+    .. autofunction:: spgrep.representation.get_projective_regular_representation
+```
+
+```{eval-rst}
+    .. autofunction:: spgrep.representation.get_intertwiner
+```
+
+```{eval-rst}
+    .. autofunction:: spgrep.representation.get_character
+```
+
+```{eval-rst}
+    .. autofunction:: spgrep.representation.project_to_irrep
+```
+
+```{eval-rst}
+    .. autofunction:: spgrep.representation.is_unitary
+```
+
+```{eval-rst}
+    .. autofunction:: spgrep.representation.is_representation
+```
+
+```{eval-rst}
+    .. autofunction:: spgrep.representation.frobenius_schur_indicator
+```
+
+```{eval-rst}
+    .. autofunction:: spgrep.representation.check_spacegroup_representation
+```
+
+```{eval-rst}
+    .. autofunction:: spgrep.representation.get_direct_product
+```
