+++ develop/_sources/api/api_group.md
@@ -1,0 +1,37 @@
+# Group
+
+```{eval-rst}
+    .. autofunction:: spgrep.group.get_cayley_table
+```
+
+```{eval-rst}
+    .. autofunction:: spgrep.group.get_identity_index
+```
+
+```{eval-rst}
+    .. autofunction:: spgrep.group.get_inverse_index
+```
+
+```{eval-rst}
+    .. autofunction:: spgrep.group.get_order
+```
+
+```{eval-rst}
+    .. autofunction:: spgrep.group.is_matrix_group
+```
+
+```{eval-rst}
+    .. autofunction:: spgrep.group.get_factor_system_from_little_group
+```
+
+```{eval-rst}
+    .. autofunction:: spgrep.group.get_little_group
+```
+
+```{eval-rst}
+    .. autofunction:: spgrep.group.check_cocycle_condition
+```
+
+```{eval-rst}
+    .. autofunction:: spgrep.group.decompose_by_maximal_space_subgroup
+```
