### 2023-05-23 #5269 @danielskatz
@lan496 - I'm the track editor who will handle the final processing for this submission. I have proofread it, and suggest one minor changes in https://github.com/spglib/spgrep/pull/86  Please merge this (unless you disagree with it), then we can proceed with acceptance and publication.
